+++ packages/plugin-hardhat/src/utils/contract-address.ts
@@ -1,0 +1,9 @@
+export type ContractAddressOrInstance = string | { address: string };
+
+export function getContractAddress(addressOrInstance: ContractAddressOrInstance): string {
+  if (typeof addressOrInstance === 'string') {
+    return addressOrInstance;
+  } else {
+    return addressOrInstance.address;
+  }
+}