+++ packages/plugin-hardhat/src/utils/index.ts
@@ -3,3 +3,4 @@ export * from './deploy-impl';
 export * from './factories';
 export * from './is-full-solc-output';
 export * from './validations';
+export * from './contract-address';